fix tests w/ Devise
The Devise::TestHelper class should only be mixed into the controller
tests.  I misread the info in the Devise README.  This is the way the
test_helper.rb should be.

Before this change, controller tests ran fine, but model tests would
error out with the error, “NoMethodError: undefined method `env' for
nil:NilClass”.