validation and tests for enrollment
basic functional tests and some model tests.  The only thing I couldn’t
get working was the function test for sections#drop.  It claims that
after the drop action the count of enrollments doesn’t decrease.  So
it’s commented out for now so the rest of the tests pass.