:+1: got the calendar invites working!
Calendar invitations should be working now (as well as cancels when you
drop a course) under all versions of Outlook.  Apparently Outlook likes
it when the ics calendar event data is in the body of the email, so we
put it there in addition to being an attachment.

Tested in Outlook 2013 on Windows 7 as well as in Outlook Web Access
and they both worked as expected.  Outlook for Mac 2011 is working too
(on version 14.1.0 at least)