updated section views and general refactoring
Kind of got down a rabbit hole with this stuff, but I think we found
our way out into a better place.  It all started with looking at
cleaning up the views for sections.  Here’s a rundown:

- Modified the User enrollments? and instructing? method so that they
can take *either* a Course or Section as their argument.  This
simplified these checks when you already have a specific section.  It
also eliminated the ‘@enrolled_section’ instance var from the
courses#show action.  Things work out a bit cleaner now.
- renamed the User#enrollment_for_course method to just
User#enrollment_for.  It can now also take either a Course or Section
as its argument.
- Updated the section#index and course#show views to highlight any
sections that you are instructing.  Also removed the “Enroll” button if
you are instructing the class.
- Removed the sections#roster action and abilities.  When looking at
the sections, the sections#show action was essentially the same thing,
so I added the list of enrollments to that view.  Renamed the ability
back to :view_enrollments.  I liked that term for the permissions
better anyway, it’s more of an action than :roster.
- Refactored the views using the course title to use a common partial.
Also partialized the “badges” that display under the title, e.g. for if
you are enrolled or instructing a course.
- ordered the sections by section number in lists
- made the section actions consistent across views