Multi-part sections for courses
Split sections into parts, so a section can have multiple parts, to
support the “multi-part workshops”.  The UI needs much work now, as
displaying a section with multiple parts is currently pretty ugly.
Functionally, it all works.  Parts are added along with a section, and
a section should have at least one part (TODO:  add a validation for
that).