worked on the section views
trying to come up with a slick way to display the available sections
for a course.  think I’ve gotten something that really cleans up a lot.

The available sections list on both the courses#show action as well as
the sections#index action both use the same ‘section’ partial, so that
DRYed things up nicely.  Also, in getting rid of the tables I was able
to play with the responsive formatting and make it look good on smaller
devices as well.

Still need to work on the “current enrollments” list and the list of
sections that you are instructing.