Moved the instructor back to to Session
To make life easier for us down the road, and in the interest of
keeping things simple, moved the instructor association back to the
Session model from the Part model.  This will make course evaluations
much easier, as the evaluation will be associated with a Section and we
won’t have to worry about individual parts.

WIP:  The ‘_section’ partial is a bit of a mess right now, I’ll get
that cleaned up here shortly.