prettied up and mobilized the calendar
Made the calendar a bit prettier, added a list-view at the bottom of
the calendar, and then used that as the mobile view, hiding the big
calendar.  Ideally, we should do something to detect a mobile browser
and not even render the big calendar - but that sounds like an issue
for another time.