user dashboard and profile/settings
moved the users#show page to be a ‘dashboard’ to give an overview of
enrollments, courses, and your report card.  This is still the show
action for a user, but is also the root when you are logged in.
Removed user information display from this page.

Added users#profile and users#settings actions to the users controller
for editing and managing user profile info and settings.