Refactored User::find_or_create and added tests
Set the User::find_or_create method to receive an options hash
rather than just a username.  The query for the user will still
use the :username as the key, but if a user needs to be created,
then this hash can contain the values with which to create the
user.

Also added tests for User::find_or_create and the sections#enroll_user
action.