Refactored enrollment emails into model hooks
Rather than sending the emails from the sections#enroll and
sections#drop actions, pull them into the Enrollment model as
after_create and after_destroy hooks.

This gives a better flow, and that logic doesn't really belong in
the Sections controller.  Also DRYs things up for when we add the
enroll_user and drop_user actions.