Fix display of section end times
The production servers are set to use UTC time, and the app is
configured to know to use the 'America/New_York' timezone.  The issue
was that when calculating the end time for a course based on the start
time + duration, it was using `Time.at (which directly uses the operating
system time) vs. using `Time.zone.at` (which applies the time zone
configured in the application).

Also switched to using `Time.current` rather than `Time.now` (Time.current is
an alias for Time.zone.now).