Added upcoming course feed; REFACTOR!
Added feeds for the upcoming courses.  With the upcoming feed being
driven by the date/time for a section part, it didn't really make sense
to me to keep it in the courses controller, kinda breaks the REST
principle.  So pulling the feed actions out into a feed controller made
more sense, with feed#index and feed#upcoming actions.

This leaves the course controller alone, and keeps things DRY and tidy.