registration email - unenroll & display name
Added a link to the registration email that a user can click to unenroll
from a course.  Since the actual sections#drop action is a DELETE
action, I added an intermediate confirmation page.  So clicking the link
from the email takes you to the confirmation page, and on that page is a
link to drop the course.  This is a safer approach.

Also changed the greeting in the email to use the display_first value
(either nickname or first name) instead of always using the first name.

One other tiny change, prior to this commit when you enrolled or dropped
a course, it would redirect you back to the page that you were on.  This
was fine in the beginning, but as the app has grown this is beginning to
get a bit cumbersome to manage.  So I think it's a fair compromise to
have these actions redirect you to the course/section page.  If it
becomes a usability issue, then we can always address later :)