Explicitly set license key from environment var
The docs say that any environment variable starting with `NEW_RELIC_`
will automatically be picked up by the agent, but it doesn't seem that
way - at least not with the license key.  So I had to manually specify
it in the newrelic.yml file as `ENV['NEW_RELIC_LICENSE_KEY']`.

Also set the app name :)