Add validations on Course.short_title
There were no validations at all on the :short_title attribute for a
Course.  Added a presence validation, as well as a length validation to
ensure a maximum length of 30 chars.

For convenince, a :before_validation callback was added to set
:short_title = :title if there is no :short_title.

As I was adding tests for these I also noticed that the :summary
attribute didn't have any validations either, so added a presence
validation on that field as well.