remove redundant `after_sign_in_path_for` method
The `after_sign_in_path_for` method in the ApplicationController was
duplicating the default Devise functionality.  The only different piece
was that we had mapped the users#show action to the named route
'dashboard', whereas Devise will automatically look for a named route
for a resource, '<resource>_root'.  This was the only place
'dashboard_path' was referenced, so removed it and renamed the route to
'user_root'.  So now upon successful authentication, Devise will
redirect to 'user_root_path'.