initial commit of wait-list feature
The basic wait-list feature.  Added a boolean attribute 'waiting' to the
Enrollment model to track whether a user is waiting or actively enrolled
in a section.  Then had to retroactively update everything that broke :)

This is just a rough commit, there's still some refactoring and cleanup
to do, but it serves the basic waiting list functionality.

- If a section is full the [Enroll] button changes to [Wait for a spot]
- The section#show view shows the Course Roster and Waiting List
- The stats#enrollment dashboard indicates the # of people waiting
- The users#show page 'current_enrollments' section indicates courses
  you are waiting on.