Don't show the time for "online" courses
"Online" courses are denoted as starting at '12:00am'.  This was
accounted for in the parts/_part partial, but not on the calendar.  To
keep things consistent, added an `online?` method to the Part model (to
indicate if a part is online or not) and then used that in the calendar
to hide the start time if it's an online part.