fix dropping users from courses in users#show
The users#show view made an incorrect assumption that the user
viewing the page was always the user that was logged in.  So if you
were an admin viewing the profile page of another user and clicked
the 'Drop Course' button, you'd get an error (or on the off-chance that
you were enrolled in the same course, then you would be dropped).

Updated the enrollment view to add some logic to use either the
sections#drop or sections#drop_user action based on if you were the
current logged in user or not.