Remove creating 'parts' in test fixtures
'parts' are dependent on a specific time value, `:starts_at`.  Creating
parts as test fixtures means that specific dates need to be used, and
some tests are (or will be) dependent on parts with specific `:starts_at`
values, relative to the current date (i.e. in the past, present, or
future).

With the `:starts_at` values hard-coded into the fixtures, as time
progresses, unless the fixtures are updated, as a date moves from the
future, to the present, to the past; certain tests may break simply
because of the fixture values.  This is not desireable!

By eliminating the fixtures, we can create Parts in the tests whenever
we need them, using date values relative to `Time.current` to achieve
testable values.  This should ensure consistent testablilty.

Once the `Part` fixtures were removed, it brought to light a few bugs
resulting from sections with no parts, so those were fixed as well.