Make Ability tests green!
When you change a behavior, you need to to make sure and update any
tests that tested the prior behavior.  In this case, we changed the
instructor ability so they can :view_enrollments for any sections,
rather than just sections they are teaching.

The Ability test was still trying to ensure instructors can only view
enrollments for sections they teach...but now it's not.