new relic deployments + rbenv-vars in dev
Trying to setup cap to notify new relic of deployments.  They even give
us a nice cap task to use.  But it wasn't working because the NR license
key is configured via an ENV var, which isn't available to cap.

We had been using the "figaro" gem to keep track of the environment
variables, but that depends on the application booting to read the YAML
file and set the environment.  When capistrano runs, the app doesn't
boot, so the env vars don't get set.

I've become really fond of rbenv-vars, it's simple and clean, and
doesn't depend on the app booting up.  Anything that runs through rbenv
will get the vars (if they are configured and visible).