Application settings feature
Adding a feature to store application settings.  This is a simple
key/value pair stored in a table.  Values are all string values.  A full
audit history isn't recorded, but it does store the username of the user
who last changed the value (along with the std timestamps).

WIP:  still need to update the form and ui for pages