Got some UserMailer tests; need to think/plan more
Built out some tests for the UserMailer.  Added a test for the
evaluation email, and then the enroll email.  In the enroll email, ran
into a bit of a difficult spot trying to write a fixture to test the
email content.  There is so much dynamic content in the email that
manually rigging it in the fixture felt brittle.  Actually, it felt
brittle in the evaluation email, but not too bad so I ignored it.  But
now, in the enroll email, there is so much more that I need to think on
this some more, as to the best way to approach testing the mailers.
What is most important to test?  Is the email body content someting that
is important to test?