Clear ActionMailer::Base.deliveries before tests
Looks like my issue was that the ActionMailer::Base.deliveries array
wasn't being cleaned out when Travis CI runs the tests.  Added a call to
ActionMailer::Base.deliveries.clear in the setup method for the
UserMailer tests (which shouldn't hurt anything, and it never hurts to
be explicit - even though the deliveries array should be cleared out
implicitly).

*fingers crossed*